[HAL][FDCAN]: Fixed not having a useful error handler
ST has two error handlers, one for "error status" and one for "error". The first
one gets some subset of the FDCAN_IR flags, while the second one is triggered by
a subset of the FDCAN_IR flags but doesn't get them in the function call; the
handler has to figure it out themselves (despite the flag mask not being
public!).

Additionally, note the `if (hfdcan->ErrorStatus != 0)`. This means that if *any*
error flag is set for *any* reason -- even just trying to read when there's no
data available (and guess what, we do that!) -- then that handler is called,
even if none of those relevant flags are set. That is dumb and we don't want
that.